+++ backend/src/main/java/com/cognizant/training/model/Memory.java
