+++ backend/src/main/java/com/cognizant/training/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.cognizant.training.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.cognizant.training.model.Product;
+
+public interface ProductRepository extends JpaRepository<Product, Long>
+{}
