+++ backend/src/main/java/com/cognizant/training/security/JwtAuthenticationEntryPoint.java
