+++ backend/src/main/resources/META-INF/additional-spring-configuration-metadata.json
